+++ jetnn/models/encoders/gnn/custom_convs/typed_ggnn.py
@@ -1,0 +1,92 @@
+from typing import Optional
+
+import torch
+import torch_geometric
+from torch import Tensor
+from torch.nn import Parameter as Param
+from torch_sparse import SparseTensor, matmul
+
+from torch_geometric.nn.conv import MessagePassing
+from torch_geometric.typing import Adj, OptTensor
+
+from torch_geometric.nn.inits import uniform
+
+
+class TypedGGNNConv(MessagePassing):
+
+    def __init__(
+        self,
+        out_channels: int,
+        num_layers: int,
+        aggr: str = "add",
+        bias: bool = True,
+        edge_dim: Optional[int] = None,
+        **kwargs,
+    ):
+        super().__init__(aggr=aggr, **kwargs)
+
+        self.out_channels = out_channels
+        self.num_layers = num_layers
+
+        self.weight = Param(Tensor(num_layers, out_channels, out_channels))
+        self.rnn = torch.nn.GRUCell(out_channels, out_channels, bias=bias)
+
+        if edge_dim is not None:
+            self.lin_edge = torch_geometric.nn.Linear(
+                edge_dim, out_channels, bias=False, weight_initializer="glorot"
+            )
+        else:
+            self.lin_edge = None
+
+        self.reset_parameters()
+
+    def reset_parameters(self):
+        uniform(self.out_channels, self.weight)
+        if self.lin_edge is not None:
+            self.lin_edge.reset_parameters()
+        self.rnn.reset_parameters()
+
+    def forward(
+        self,
+        x: Tensor,
+        edge_index: Adj,
+        edge_weight: OptTensor = None,
+        edge_attr: OptTensor = None,
+    ) -> Tensor:
+        """"""
+        if x.size(-1) > self.out_channels:
+            raise ValueError(
+                "The number of input channels is not allowed to "
+                "be larger than the number of output channels"
+            )
+
+        if x.size(-1) < self.out_channels:
+            zero = x.new_zeros(x.size(0), self.out_channels - x.size(-1))
+            x = torch.cat([x, zero], dim=1)
+
+        for i in range(self.num_layers):
+            m = torch.matmul(x, self.weight[i])
+            # propagate_type: (x: Tensor, edge_weight: OptTensor)
+            m = self.propagate(
+                edge_index, x=m, edge_weight=edge_weight, edge_attr=edge_attr, size=None
+            )
+            x = self.rnn(m, x)
+
+        return x
+
+    def message(self, x_j: Tensor, edge_weight: OptTensor, edge_attr: OptTensor):
+        if edge_attr is not None:
+            if edge_attr.dim() == 1:
+                edge_attr = edge_attr.view(-1, 1)
+            edge_attr = self.lin_edge(edge_attr)
+            x_j += edge_attr
+        return x_j if edge_weight is None else edge_weight.view(-1, 1) * x_j
+
+    def message_and_aggregate(self, adj_t: SparseTensor, x: Tensor) -> Tensor:
+        return matmul(adj_t, x, reduce=self.aggr)
+
+    def __repr__(self) -> str:
+        return (
+            f"{self.__class__.__name__}({self.out_channels}, "
+            f"num_layers={self.num_layers})"
+        )
